after svg

+++ demo/src/main/java/project/demo/Dot.java
@@ -5,19 +5,19 @@ import java.io.IOException;
 
 public class Dot {
 
-    private final int x;
+    private final double x;
     private final double y;
     private final double r;
     private boolean status;
 
-    public Dot(int x, double y, double r) throws IOException {
+    public Dot(double x, double y, double r) throws IOException {
         this.x = x;
         this.y = y;
         this.r = r;
 
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
